start modularisation and breaking down the monofile
